Rusty Skylines Alpha v1.7.9a May 13th 23:41
[Major]
Fixed 2 big UI bugs!
1. Each AP template was added to the layers by every single AP reference, even if they didn't reference it!
Such a strange and pretty STUPID error I did 2 months ago, but wtf for real??

2. The AP reference is stored as an element in the elements list of each layer.
During Cache initialization, I check if the length of this list in a layer has changed compared to its cache counterpart.
I did not include the APs into this cache list, so the lengths were NEVER the same and the cache was reinitialized every frame, leading to that disgusting flickering I saw in the Settings menu all the time!
The flickering is gone now. I included the AP Reference in the cache list, but it serves no function there, it's just another enum field.

I added a schedule.rs file and hooked the economy and schedule structs up to the new CityState struct, which will hold per world stuff, like money and schedule and statistics and probably more!

[Minor]
Added a Residential Zoning Indicator, WIP as hell, and also it indicates the bugginess of my WIP residential zoning demand calculations!

All zoning demand values for the closest district are now being set as variables, and are accessible to the UI as that!

Fixed the on:always and on:a flags on UI actions!

OH! I added a money indicator and in-game Time indicator!

Made it clearer in the Terrain Edits who requested an edit. Like for example a building requested that the terrain underneath be flattened.

The AP variable is now ap_var instead of setting, because now you can reference a variable instead of only settings, which expands capability DRAMATICALLY sometimes and actually yes, and also I dramatize!

I quartered the individual page size for terrain from for example 256MB for just vertices per page, to 64MB per page. Still a lot, but less risk of fragmentation ruining dozens of millions of vertices. (Also quartered indices!)

Refactored some Terrain Edit Types to use the height from WorldPos points that literally already exist instead of some weird heights.
God, I hate AI for stuff like this... Fast as hell but has such oversights.

***
Anders Hellmann with one N?
Btw I printed AMOGUS 😂😂😂😂 I am going crazy I belong in a mental asy- no...

+++ src/renderer/general_mesh_arena.rs
@@ -91,7 +91,7 @@ impl MeshPage {
         );
 
         Some(GpuChunkHandle {
-            page: page_index as u8,
+            page: page_index,
             base_vertex: (v_off / request.vertex_align) as i32,
             first_index_above: (i_off / 4) as u32,
             index_count_above: scratch.indices.len() as u32,
